[io.metrics] Fix MetricsService breaking REST API #10640
