+++ bundles/org.openhab.io.metrics/src/main/java/org/openhab/io/metrics/MetricsRestController.java
@@ -12,7 +12,10 @@
  */
 package org.openhab.io.metrics;
 
+import java.util.HashSet;
+import java.util.Map;
 import java.util.Objects;
+import java.util.Set;
 
 import javax.annotation.security.RolesAllowed;
 import javax.ws.rs.GET;
@@ -23,8 +26,14 @@ import javax.ws.rs.core.MediaType;
 import org.eclipse.jdt.annotation.NonNullByDefault;
 import org.eclipse.jdt.annotation.Nullable;
 import org.openhab.core.auth.Role;
+import org.openhab.core.config.core.ConfigurableService;
+import org.openhab.core.config.core.Configuration;
 import org.openhab.core.io.monitor.MeterRegistryProvider;
+import org.openhab.core.io.rest.RESTConstants;
+import org.openhab.io.metrics.exporters.InfluxMetricsExporter;
+import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Modified;
 import org.osgi.service.component.annotations.Reference;
 import org.osgi.service.jaxrs.whiteboard.JaxrsWhiteboardConstants;
 import org.osgi.service.jaxrs.whiteboard.propertytypes.JSONRequired;
@@ -43,24 +52,27 @@ import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.tags.Tag;
 
 /**
- * * The {@link MetricsRestController} class implements the REST endpoints for all pull based monitoring systems.
+ * The {@link MetricsRestController} class implements the REST endpoints for all pull based monitoring systems.
  *
  * @author Robert Bach - Initial contribution
  */
 @Component(immediate = true, service = MetricsRestController.class)
 @JaxrsResource
-@JaxrsApplicationSelect("(" + JaxrsWhiteboardConstants.JAX_RS_NAME + "=" + MetricsService.METRICS_APP_NAME + ")")
-@Path("")
+@JaxrsApplicationSelect("(" + JaxrsWhiteboardConstants.JAX_RS_NAME + "=" + RESTConstants.JAX_RS_NAME + ")")
+@Path(MetricsRestController.PATH_METRICS)
 @JSONRequired
 @RolesAllowed({ Role.USER, Role.ADMIN })
 @Tag(name = MetricsRestController.PATH_METRICS)
 @NonNullByDefault
+@ConfigurableService(category = "io", label = "Metrics service", description_uri = "io:metrics")
 public class MetricsRestController {
     private final Logger logger = LoggerFactory.getLogger(MetricsRestController.class);
     public static final String PATH_METRICS = "metrics";
     private @Nullable CompositeMeterRegistry meterRegistry = null;
     private final PrometheusMeterRegistry prometheusMeterRegistry = new PrometheusMeterRegistry(
             PrometheusConfig.DEFAULT);
+    private final Set<MetricsExporter> metricsExporters = new HashSet<>();
+    private @Nullable MetricsConfiguration config;
 
     @GET
     @Path("/prometheus")
@@ -79,5 +91,29 @@ public class MetricsRestController {
         meterRegistry = meterRegistryProvider.getOHMeterRegistry();
         Objects.requireNonNull(meterRegistry).add(prometheusMeterRegistry);
         logger.debug("Core metrics registry retrieved and Prometheus registry added successfully.");
+        updateMeterRegistry();
+    }
+
+    @Activate
+    protected void activate(Map<@Nullable String, @Nullable Object> configuration) {
+        logger.info("Metrics service activated.");
+        metricsExporters.add(new InfluxMetricsExporter());
+        updateConfig(configuration);
+        updateMeterRegistry();
+    }
+
+    @Modified
+    protected synchronized void modified(Map<@Nullable String, @Nullable Object> configuration) {
+        updateConfig(configuration);
+    }
+
+    private void updateConfig(@Nullable Map<@Nullable String, @Nullable Object> configuration) {
+        this.config = new Configuration(configuration).as(MetricsConfiguration.class);
+        logger.debug("Configuration: {}", this.config);
+        this.metricsExporters.forEach(e -> e.updateExporterState(config));
+    }
+
+    private void updateMeterRegistry() {
+        this.metricsExporters.forEach(e -> e.setMeterRegistry(meterRegistry));
     }
 }
